Migrate Mutation.updateSchemaVersionStatus
